requirements.txt
python-telegram-bot==22.5
httpx>=0.28
Pillow>=11.0
python-dotenv>=1.0
